requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
seaborn
keras
pillow
tensorflow
tensorflow_hub
